Edit PDF icon; dial down demo task; return to tasks.py
