+++ web/main/tasks.py
@@ -1,7 +1,7 @@
-# Generate a PDF via playwright
-from pathlib import Path
 import logging
+from pathlib import Path
 import tempfile
+from time import sleep
 
 from celery import shared_task
 from playwright.sync_api import sync_playwright, expect, Page
@@ -45,6 +45,24 @@ def generate_pdf(
 
     logger.info(f"Wrote output to {output_file}")
     return output_file
+
+
+@shared_task
+def demo_scheduled_task(pause_for_seconds=0):
+    """
+    A demo task, scheduled to run periodically in dev. To see it in action,
+    set CELERY_TASK_ALWAYS_EAGER = False in setting.py before running `fab run`.
+
+    >>> caplog = getfixture('caplog')
+    >>> settings = getfixture('settings')
+    >>> settings.CELERY_TASK_ALWAYS_EAGER = True
+    >>> with caplog.at_level(logging.DEBUG):
+    ...     _ = demo_scheduled_task.apply_async(kwargs={'pause_for_seconds': 1})
+    >>> assert 'Celerybeat is working!' in caplog.text
+    """
+    if pause_for_seconds:
+        sleep(pause_for_seconds)
+    return "Celerybeat is working!"
 
 
 if __name__ == "__main__":
